Add subscriptions to custom highlights and centrally manage highlight colors
To start off with, I'm sorry for this commit containing two different
changes that should really be split into two commits. However, it would
be a pain to separate these two changes afterwards so I'm rolling with
it as is.

As a first change, HighlightPhrases now store a shared_ptr to the
highlight color instead of the color. This allows Message objects to
also store a pointer to the highlight color instead of having to copy
the QColor to every message. This should also allow highlight colors to
be dynamically updated after the setting is changed (also it does not
work in this commit yet).

Furthermore, the ColorProvider class has been introduced. It provides a
central place to manage the preset highlights (self highlight, whispers
and subscriptions). On initialization, the colors for the preset
highlights are set to a default or to the value read from settings.