Added possibility to change image uploader's form body with Enviroment Variables (#1709)
Updated the Tooltip to diplay the name of the artist for non global emotes, for both bttv and ffz

Changed variable to author to keep it consistent, removed typo that kept the program from compiling

auto format accidentally messed up the file format when I made a typo on the first commit, I didn't check 'git diff'.  I reverted the changes to master for the file affected and redid my portion of the code

bttv doesn't have 'name' datafield for channel emotes, when author.string is empty then 'Channel Emote' is displayed, otherwise it shows the author of the emote.

Formated the code I changed

formated changes