Added create clip when channel is live
While I was at it, I needed to use TwitchChannel property which was already dynamic_cast'ed in few places, so I turned it into a variable and refactored some very ugly 2019 code. I'm sorry for making unnecessary changes, but I couldn't resist it. That duplicated if statement was so wrong.