fix: properly handle --channels argument
With this commit, the --channels argument is respected again. The layout
parsing was refactored from the previous version using JSON generation
into directly building a `WindowLayout`.

In order to use the "correct" size for the temporary window, we read the
window layout file and get the saved geometry of the serialized main
window. This required some changes in the initialization order of the
`Paths` singleton, and making `WINDOW_LAYOUT_FILENAME` a static member
of `WinodwManager` instead of a variable in the unnamed namespace.