fix: properly handle --help and --version arguments
Prior to this commit, a logging category was used to output the `--help`
and `--version` texts. These categories are not meant to be used for
"nominal" output however: They only output messages above their
respective logging thresholds which is `QtWarningMsg` for non-debug
builds.

With this patch, the default logging category is used instead.