fix: no longer crash on clip creation in "special" channels
Clip creation via the `Alt+X` shortcut failed for the `/mentions` and
`/whispers` channel. This happened because `Channel::isTwitchChannel`
returns `true` for these channels. (Internally, the types are
`Channel::Type::TwitchMentions` and `Channel::Type::TwitchWhispers` in
anticipation of supporting more platforms in the future.)
This lead to a false assumption in the clip creation logic: Not all
channels that display "Twitch content" are able to create clips.

The employed fix checks the channel type in more detail. Clip creation
is only allowed if the channel is `Channel::Type::Twitch` or
`Channel::Type::TwitchWatching`. This makes sure that there always is an
associated channel to generate a clip from.

Additional note: Clip creation did not crash when initiated via the
`/clip` command. This is due to the fact that
`CommandController::execCommand` makes additional checks regarding the
channel context before executing the command.