Fix write connection reconnection issues
Chatterino currently does not handle disconnects of the write connection, which
causes messages to remain buffered until you manually reconnect, after which it
will send all messages at once.

This patch addresses the lack of reconnection specifically, and attempts to
refactor some of the reconnection logic to make it more robust. Backoff behavior
is introduced to *both* read/write cons to avoid reconnection loops.

To replicate the issue, simply lose your (write) connection to TMI. This can be
"simulated" using something like:

    tcpkill -i eth0 host 44.226.36.141 or host 34.217.198.238 or host 100.20.159.232

An alternative/secondary approach would be to detect failures of the underlying
`sendRaw` method (as it returns false if the message can't be sent to the
connection), but better monitoring of the connection state should be sufficient.

The way the code is setup currently doesn't really make it easy to implement
this "cleanly". E.g. calls `initializeConnection` are skipped on reconnect, but
since we're reconnecting with the same parameters, this should have little
actual consequence.