fix: provide better UX for HTML color picker edit
Before this change, it was checked whether the text of the HTML color
picker line edit represents a valid color. Since the HTML color syntax
allows the format `#RGB` (see [1]), this lead to the lambda setting the
color edit to the full format string (`#RRGGBB`) whenever three
characters have been entered.
Due to the QValidator set up for the line edit, attempts to change the
text again would be rejected.

With this change, validity of the color is only checked when the
`QLineEdit::editingFinished` signal is emitted. (At the time of writing,
this corresponds to "Return" being pressed or a focus loss of the line
edit.) Since the signal is only emitted when the validator approves the
input (see [2] for more information), this will never lead to the color
picker being in an inconsistent state.

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/color_value#syntax_2
[2]: https://doc.qt.io/qt-5/qlineedit.html#editingFinished