ref: simplify spawning Streamlink process on Flatpak
As suggested in written communication with @ilyazzz, we now
incrementally build the Streamlink process by specifying the correct
path and adding the required arguments at different locations in the
code.

The correct path is `flatpak-spawn` when Chatterino is running as a
Flatpak or the direct path to the Streamlink executable otherwise.
Arguments are chosen depending on the context Streamlink is called from,
e.g. for fetching quality options or for opening a stream.

The `getStreamlinkProgram()` has been removed and its functionality has
been integrated in `createStreamlinkProcess()` instead.