fix: re-add date of build to "About" page
The date of build has previously been provided via a define supplied by
the CI runner. During the migration to GitHub Actions, this seems to
have been missed.

This fix relies on CMake generating and providing the timestamp itself,
and passing it to the binary. Note that the timestamp will be of the
date when the **build files** were generated, i.e., not when the project
was compiled! For CI runs, however, the date of build file generation
should be consistent with the date of compilation so this approximation
is "good enough" for our purpose.