Only change live status for IDs present in batch
This mitigates a side-effect spotted in pajlada's review [0] which was caused by iterating over a wrong range - within the `getHelix()->fetchStreams(IDs, ...)` request call, only channels present in IDs should be "touched", not all of them.
This seems like a perhaps less algorithmicly-complex solution than the one pointed out in [0] (which I've created in [1] and pushed to [2] branch)

And while I was at it, I've batch splitter helper use generic type template and renamed it + its parameter to reflect its scope being a bit more universal, rather than dedicated just to 1 scenario (which it still is, but should be changed in the future).

[0]: https://github.com/Chatterino/chatterino2/pull/3767/files#pullrequestreview-983862675
[1]: https://github.com/Chatterino/chatterino2/commit/0e208662057058ca15c9f68fda0fcaf04f1fe8fb
[2]: zneix/fix/bulk-stream-fetch-set-offline-alt1