Corrected the way we parse "list" tags in PRIVMSG
Why was the logic change in parsing "badges" and "badge-info" made/needed:

It's been discovered that elements of the comma-separated list (taglist) in "badge-info" tag can have slashes in their values.
For example, "foo/bar/baz,tri/hard" can be a valid badge tag
In that case, valid map content should be:
    {"foo": "bar/baz", "tri": "hard"}
With old logic of just splitting elements of the taglist by (every) slash it would be instead:
    {"foo": "bar" "baz", "tri": "hard" }
Here, the "foo" element of the taglist would be considered invalid and skipped (because the amount of substrings after `QString::split` operation was not equal to 2).
Now, the lack of slash can indicate that the tag is invalid and should be skipped instead of checking for amount of substring being non-equal to 2.
And to ensure we parse taglist's elements properly, a `QString::section` is used which can return a substring of elements before and after a specified occurence of the delimeter, similarly to GNU Coreutils' cut command:
    zneix@uds ~$ echo "foo/bar/baz" | cut -d/ -f1
    foo
    zneix@uds ~$ echo "foo/bar/baz" | cut -d/ -f2-
    bar/baz

Details about codebase changes:

In the process a lot of duplicate code was removed to have the logic responsible for this in one place only. For that to work, some type changes were required.
I've also decided in this scenario it's better to use `std::unordered_map` over `std::map` - perhaps it's convinient to use `QHash` at this point, but I'm not sure so I just left it as is.

Changing the default constructor for `Badge` class in TwitchBadges.cpp made things easier to handle, making `SharedMessageBuilder::parseTagList` work for parsing both `badge-info` and `badges` tags in PRIVMSG messages, without having to duplicate the same function just for the sake of returning `std::vector<Badge>` instead of `std::(unordered_)map<QString, QString>`. This comes with a little compromise in TwitchMessageBuilder.cpp, where a helper vector is declared and set as `this->message().badges` - perhaps this can be mitigated somehow.