ref: optimistically load historic messages
Prior to this commit, historical messages for a channel were loaded when
the ROOMSTATE message for the channel was received. Thus, loading of
historical messages was limited by our IRC join limit although the
recent messages provider we use may have an entirely different rate
limit!

This commit decouples receiving the ROOMSTATE message for a channel from
requesting its historical messages. This allows us to
    1. speed up loading times if the recent messages service enforces a
       laxer rate limit than Twitch IRC, and
    2. adhere to a stricter rate limit if the recent message provider
       requires one.

Additionally, this commit changes the "fetching priority" of channels.
Previously, historical messages were fetched in alphabetical order. With
this commit, channels are fetched in "tab order", i.e., channels that
appear in tabs earlier in the layout will be loaded first. I believe
this will further improve UX, assuming that channels "more important" to
the user appear earlier.