Refactor pagination logic
Instead of having a shared method to handle the pagination logic, this
adds a private `fetchChatters` method that expects the caller to
provide all parameters that would be necessary to make the request.
This method can then be used from the public-facing `getChatters` method
in a recursive manner, while also keeping the error handling code right
in that function.

With this change, the success result type has been changed to the
`HelixChatters` type (the same used in the private method). By doing
this, we can provide a consistent & full API for any users of the
`getChatters` API and not have to provide a separate API for the
`/chatters` function.

Other important changes that came with this was also avoiding the use of
raw pointers, the previous code had difficult-to-track pointers & memory
leaks for every call that was made (chatterList and page).

One potential unbenchmarked downside to the refactored implementation is
that we make more sets of users instead of adding them straight onto the
same one. This is a concession that is worth it in my opinion as it
makes the implementation easier to read. If the code was used in some
hot spot we can benchmark and reconsider this.

Another bonus change added to this feature is that we fetch 1000
chatters per page using the `first` request query parameter, making us
use a lot fewer API requests. This means we'll run less functions
recursively and worry less about crashing our stack.