feat: Add a basic implementation of shared chat profile pictures
The biggest problems with this implementation, and the ones which need
to be fixed ASAP are:
* It stores the URL of the profile picture on the TwitchUser class
for each user, which may or may not increase the ram usage.
* It uses a hard-coded replacement of the 300x300 size which is, while
seemingly consistent across channels, just looks very bad.
* Quite a bit of code duplication.
* Some inconsistent variable naming.

Besides that, from the minimal amount of testing I've done, it seems
to be pretty usable.