various fixes
Close open EventSub connections on exit

Helix function to create subscriptions now accepts a SubscriptionRequest

Added a debugger for SubscriptionRequest

Added a `SubscriptionHandle` that can/will be able to take care of unsubscriptions from topics once no one is interested in them anymore. They're meant to work like scoped signal connections, where the dtor of the handle is what spits lowers the reference count of the "topic".

There's a bad subscription for `channel.ban` in TwitchChannel atm, but it's being used to test subscription handles and errors in permissions