fix: typo in color input selector. changed from red to blue issue #5979
Changed the label for "Blue" back to Blue as it was incorrectly changed to "Red". This bug was introduced in commit 78a7ebb9f952d4de8fb69b95a786e570e96bdf1c.

To replicate this bug, open any of the color picker windows. One example is the "Highlights" menu.
![image](https://github.com/user-attachments/assets/1a6f8d1f-57cf-44dd-82d9-4e56f4234af6)