feat(SettingWidget): color button
This differs from the GeneralPageView color button implementation in
that we don't listen to the signal from the ColorPickerDialog to
update the ColorButton color, rather we use that to update the setting
which in turn updates the ColorButton color.

This means that pressing "Cancel" on the Settings dialog will now
correctly revert the ColorButton color to the pre-change color.

It does mean an extra conversion, but it doesn't matter here