Add `setQuitting` method to eventsub::IController
This tells the eventsub controller that we are in the process of
quitting, meaning any future network requests that might want to be made
can just be skipped.

This means we might leave websocket connections in a worse state, but I
_think_ that's fine - Twitch should just kill those eventsub
subscriptions for us. If Twitch doesn't do that, we will/would have
problems with the app crashing too.