refactor(SelectChannelDialog): remove twitch tab (and other stuff)
We no longer use a Notebook for this, this means the layout we have will
autosize appropriately.
To make it autosize on slightly larger font sizes I've done manual word
wrapping on the labels instead of letting them word wrap themselves.

If the user resizes the dialog after opening it, things might look a
little bit more goofy than before, but the default size it opens as
should be better.

Something with how I add labels now (not using the layout creator)
    messed up the focus chain. I've fixed this by checking 2 steps ahead
    in the nextInFocuschain and previousInFocusChain.

the way we want this to work, we cannot use Qt's own focus chain logic
since the QTextEdit that belongs with the "Channel" radiobox will
always be ignored since it's not visible