nit: remove `MarkdownLabel::initializeMarkdownDocument` method
We should in general only use this pattern when the constructor is
either:
a) really big, so splitting it up into functions makes it clear what the
constructor does at a glance
b) follows a pattern shared with a lot of other similar classes, where
it would be easier to at-a-glance know what the constructor does in
addition to the called method (e.g. an `initUI` method)
c) the same logic needs to be reused in other places (e.g. an
    `updateText` method)